made compatible with Healpix.jl v3.0.0

+++ examples/data/gen_pysm_maps.py
@@ -8,7 +8,7 @@ for i in range (0,10):
     map = sky.get_emission(nu[i] * u.GHz)
     map_equ = pysm3.apply_smoothing_and_coord_transform(map, rot=hp.Rotator(coord="GC"))
     # RING format
-    hp.write_map("map_equ_{:03d}ghz_r8_uKcmb.fits".format(nu[i]), map_equ, coord="C")
+    hp.write_map("map_equ_{:03d}ghz_r8_uKcmb.fits".format(nu[i]), map_equ, coord="C", dtype=np.float32)
     # NESTED format
     map_nested = hp.reorder(map_equ, r2n=True)
-    hp.write_map("map_equ_{:03d}ghz_r8_nested_uKcmb.fits".format(nu[i]), map_nested, coord="C", nest=True)
+    hp.write_map("map_equ_{:03d}ghz_r8_nested_uKcmb.fits".format(nu[i]), map_nested, coord="C", nest=True, dtype=np.float32)
